The buildASTSchema function generates a "Schema" object from an AST that represents a schema doc. The graphql-go module has support for parsing a schema SDL to an AST, but does not have a way to then get a Schema object.
We need the Schema object for pruning because we need the type info
from Schema objects in order to run the visitor through the query
document with the VisitWithTypeInfo option.

This is a large change, but the logic is mostly just converting node
objects in the schema AST (which have go types from 'ast' package)
to corresponding "type" objects (which have go types from the
'graphql' package). These "type" objects are what make up the Schema.